Separated Commands and Options into traits
