+++ RoboFile.php
@@ -1,5 +1,4 @@
 <?php
-
 use Robo\Tasks;
 
 class Robofile extends Tasks
